visualization/requirements.txt
psycopg2
matplotlib
python-dotenv